Lots of glue
Hooking everything up. Score is broadcast in the ScoreChange event.
Time is broadcast in the TimeChange event. GameStart event called on
ManagerStart and inside of Play in the AudioManager to reset the clock.
Also added a distance traveled script to avoid messing up anything in
your worm scripts.

I’m a little tired right now, but everything should just work at this
point. (systems anyways, all the actual events need to be added in
still).